Vision pipeline corrections and tests
Refactored the vision pipeline for gear drop detection so the entire
codebase is in a separate/testable package independent of WPILIB. Added
a GripPipelineTest to operate on a known image and product the same
output as the GRIP GUI.  Added workaround in the Generated code to deal
with https://github.com/WPIRoboticsProjects/GRIP/issues/811 which I
logged.

DO NOT regenerate the GripPipeline from GRIP until this issue is fixed
or the blob detection will break!